+++ tests/testthat/test-FLQuant.R
@@ -1,0 +1,8 @@
+# test-FLQuant.R - DESC
+# /test-FLQuant.R
+
+# Copyright Iago MOSQUEIRA (WMR), 2020
+# Author: Iago MOSQUEIRA (WMR) <iago.mosqueira@wur.nl>
+#
+# Distributed under the terms of the EUPL-1.2
+